[iPhoneXMPP] Added background support & message alerts.
Configured iPhoneXMPP to run in the background.

Added message and new buddy notifications. iPhoneXMPP now uses UIAlertView or UILocalNotications to notify the user when a message or buddy request arrives.  There is still no way to accept a buddy request or to send a message.

Added XMPPvCardTempModule to the app.  The XMPPvCardAvatarModule needs to be ported to v3, before we can add photos to the roster display.

--HG--
branch : v3