[XMPPRosterUserCoreDataStorage] Add transient caching and sectionName for roster sorting.
Changed jid and section to on-demand cached transients.

Added sectionName as an on-demand cached transient.  sectionName allows for sorting by the first character of the displayName.  This can be used to section the roster by displayName, instead of by availability (sectionNum).  sectionName is also useful for UITableView indexes.

--HG--
branch : v3