added explicitly set locale on NSDateFormatters to bypass Apple bug.
There is a know NSDateFormatter bug the causes date parse to fail on this document in some cases. The bug happens in some cases when the user sets a different locale other than 'en_US'. The bug causes NSDateFormatter to expect an am/pm even when sent to parse 24hr strings. This is easily solved by explicitly setting the locale of the NSDateFormatter.