Removed unnecessary code and logging.
XMPPMessageArchiving is the only extension which writes to the debug log that it is activated itself. I find this very annoying in the log, since when I want to check which plugins got activated I would use the delegate method and I don't want a plugin to do the logging itself. (Especially since you (nearly) can't disable it when using CocoaPods.)